requirements.txt
streamlit>=1.24.0
Pillow>=9.0.0
colorthief>=0.2.1
openai>=1.12.0
requests>=2.28.0